requirements.txt
pandas
numpy
Pillow
jsonschema
argparse
opencv-python
matplotlib
bpy
